Внесены незначительные правки правки
