+++ task/src/main/java/com/sinx/task/di/TaskComponent.kt
@@ -1,9 +1,7 @@
 package com.sinx.task.di
 
-import android.content.Context
 import com.sinx.core.Feature
 import com.sinx.task.TaskListFragment
-import dagger.BindsInstance
 import dagger.Component
 
 @Feature
