--- task/src/main/java/com/sinx/task/di/TaskComponentViewModel.kt
@@ -1,8 +1,0 @@
-package com.sinx.task.di
-
-import androidx.lifecycle.ViewModel
-
-//class TaskComponentViewModel : ViewModel() {
-//    val newDetailComponent =
-//        DaggerTaskComponent.builder().deps(TaskDepsProvider.deps).build()
-//}
